use utils.is_changed instead depend.on_changed

+++ xmake/modules/skr/install.lua
@@ -23,7 +23,7 @@ function install_tool(opt)
 
     -- extract package
     local package_file_name = path.filename(package_path)
-    depend.on_changed(function ()
+    if opt.force or utils.is_changed(utils.depend_file("install/tools/"..package_file_name), {package_path}) then
         _log(opt, "tool.install", "path \"%s\"", package_path)
         archive.extract(package_path, utils.install_dir_tool())
 
@@ -32,12 +32,7 @@ function install_tool(opt)
             utils.redirect_fenv(opt.after_install, install_tool)
             opt.after_install()
         end
-    end, {
-        dependfile = utils.depend_file("install/tools/"..package_file_name),
-        lastmtime = os.mtime(package_path),
-        files = {package_path},
-        changed = opt.force,
-    })
+    end
 end
 function install_resources(opt)
     local package_path = utils.find_download_package_sdk(opt.name)
@@ -50,21 +45,19 @@ function install_resources(opt)
     local extract_exist = os.isdir(temp_dir)
 
     -- extract
-    depend.on_changed(function ()
+    if not extract_exist or opt.force or
+       utils.is_changed(utils.depend_file("install/resources/extract/"..depend_file_name), {package_path}) then
+        
         _log(opt, "resource.extract", "temp_dir \"%s\"", temp_dir)
         if extract_exist then
             os.rmdir(temp_dir)
         end
         archive.extract(package_path, temp_dir)
-    end, {
-        dependfile = utils.depend_file("install/resources/extract/"..depend_file_name),
-        lastmtime = os.mtime(package_path),
-        files = { package_path },
-        changed = not extract_exist or opt.force,
-    })
+    end
 
     -- copy
-    depend.on_changed(function ()
+    if not extract_exist or opt.force
+       or not utils.is_changed(utils.depend_file("install/resources/copy/"..depend_file_name), {package_path}) then
         _log(opt, "resource.copy", "path \"%s\"", package_path)
         if not os.isdir(opt.out_dir) then
             os.mkdir(opt.out_dir)
@@ -76,12 +69,7 @@ function install_resources(opt)
             utils.redirect_fenv(opt.after_install, install_resources)
             opt.after_install()
         end
-    end, {
-        dependfile = utils.depend_file("install/resources/copy/"..depend_file_name),
-        lastmtime = os.mtime(package_path),
-        files = { package_path },
-        changed = not extract_exist or opt.force,
-    })
+    end
 end
 function install_sdk(opt)
     local package_path = utils.find_download_package_sdk(opt.name)
@@ -93,21 +81,18 @@ function install_sdk(opt)
     local extract_exist = os.isdir(temp_dir)
 
     -- extract
-    depend.on_changed(function ()
+    if not extract_exist or opt.force or
+       utils.is_changed(utils.depend_file("install/sdk/extract/"..package_file_name), { package_path }) then
         _log(opt, "sdk.extract", "temp_dir \"%s\"", temp_dir)
         if extract_exist then
             os.rmdir(temp_dir)
         end
         archive.extract(package_path, temp_dir)
-    end, {
-        dependfile = utils.depend_file("install/sdk/extract/"..package_file_name),
-        lastmtime = os.mtime(package_path),
-        files = { package_path },
-        changed = not extract_exist or opt.force,
-    })
+    end
 
     -- copy
-    depend.on_changed(function ()
+    if not extract_exist or opt.force or
+       utils.is_changed(utils.depend_file("install/sdk/copy/"..package_file_name), { package_path }) then
         _log(opt, "sdk.copy", "path \"%s\"", package_path)
         if not os.isdir(opt.out_dir) then
             os.mkdir(opt.out_dir)
@@ -125,12 +110,7 @@ function install_sdk(opt)
             utils.redirect_fenv(opt.after_install, install_sdk)
             opt.after_install()
         end
-    end, {
-        dependfile = utils.depend_file("install/sdk/copy/"..package_file_name),
-        lastmtime = os.mtime(package_path),
-        files = { package_path },
-        changed = not extract_exist or opt.force,
-    })
+    end
 end
 function install_file(opt)
     local file_path = utils.find_download_file(opt.name)
@@ -139,7 +119,8 @@ function install_file(opt)
     end
     local depend_file_name = opt.name..(opt.trigger_target and "_"..opt.trigger_target or "")
 
-    depend.on_changed(function ()
+    if opt.force or
+       utils.is_changed(utils.depend_file("install/files/"..depend_file_name), { file_path }) then
         _log(opt, "download_file.copy", "path \"%s\"", file_path)
         if not os.isdir(opt.out_dir) then
             os.mkdir(opt.out_dir)
@@ -151,12 +132,7 @@ function install_file(opt)
             utils.redirect_fenv(opt.after_install, install_file)
             opt.after_install()
         end
-    end, {
-        dependfile = utils.depend_file("install/files/"..depend_file_name),
-        lastmtime = os.mtime(file_path),
-        files = { file_path },
-        changed = opt.force,
-    })
+    end
 end
 function copy_files(opt)
     -- check opt
@@ -175,7 +151,8 @@ function copy_files(opt)
     end
 
     -- copy
-    depend.on_changed(function ()
+    if opt.force or
+       utils.is_changed(utils.depend_file("install/copy_files/"..opt.name), files) then
         _log(opt, "file.copy", "%d files", #files)
         if not os.isdir(opt.out_dir) then
             os.mkdir(opt.out_dir)
@@ -183,22 +160,15 @@ function copy_files(opt)
         for _, file in ipairs(files) do
             os.cp(file, opt.out_dir, {rootdir = opt.root_dir})
         end
-    end, {
-        dependfile = utils.depend_file("install/copy_files/"..opt.name),
-        files = files,
-        changed = opt.force,
-    })
+    end
 end
 function install_custom(opt)
-    depend.on_changed(function ()
+    if opt.force or
+       utils.is_changed(utils.depend_file("install/custom/"..opt.name), {path.join(os.scriptdir(), "install.lua")}) then -- TODO. use flag file
         _log(opt, "custom")
         utils.redirect_fenv(opt.func, install_custom)
         opt.func()
-    end, {
-        dependfile = utils.depend_file("install/custom/"..opt.name),
-        files = { os.scriptdir() }, -- just for one time run
-        changed = opt.force,
-    })
+    end
 end
 -- opt see xmake/rules/install.lua
 function install_from_kind(kind, opt)
